+++ src/array.rs
@@ -18,7 +18,7 @@ use std::ptr;
 use std::ops::*;
 
 /// An array containing elements of type `Element`
-pub trait Array1<Element: Copy>: Index<uint, Output=Element> + IndexMut<uint, Output=Element> {
+pub trait Array1<Element: Copy>: Index<usize, Output=Element> + IndexMut<usize, Output=Element> {
     /// Get the pointer to the first element of the array.
     fn ptr<'a>(&'a self) -> &'a Element {
         &(*self)[0]
@@ -31,14 +31,14 @@ pub trait Array1<Element: Copy>: Index<uint, Output=Element> + IndexMut<uint, Ou
 
     /// Swap the elements at indices `i` and `j` in-place.
     #[inline]
-    fn swap_elems(&mut self, i: uint, j: uint) {
+    fn swap_elems(&mut self, i: usize, j: usize) {
         // Yeah, ok borrow checker – I know what I'm doing here
         unsafe { ptr::swap(&mut (*self)[i], &mut (*self)[j]) };
     }
 
     /// Replace an element in the array.
     #[inline]
-    fn replace_elem(&mut self, i: uint, src: Element) -> Element {
+    fn replace_elem(&mut self, i: usize, src: Element) -> Element {
         mem::replace(&mut (*self)[i], src)
     }
 
@@ -49,7 +49,7 @@ pub trait Array1<Element: Copy>: Index<uint, Output=Element> + IndexMut<uint, Ou
 
 /// A column-major array
 pub trait Array2<Column: Array1<Element>+'static, Row: Array1<Element>, Element: Copy>:
-        Index<uint, Output=Column> + IndexMut<uint, Output=Column> {
+        Index<usize, Output=Column> + IndexMut<usize, Output=Column> {
     /// Get the pointer to the first element of the array.
     fn ptr<'a>(&'a self) -> &'a Element {
         &(*self)[0][0]
@@ -62,25 +62,25 @@ pub trait Array2<Column: Array1<Element>+'static, Row: Array1<Element>, Element:
 
     /// Swap two columns of this array.
     #[inline]
-    fn swap_cols(&mut self, a: uint, b: uint) {
+    fn swap_cols(&mut self, a: usize, b: usize) {
         unsafe { ptr::swap(&mut (*self)[a], &mut (*self)[b]) };
     }
 
     /// Replace a column in the array.
     #[inline]
-    fn replace_col(&mut self, c: uint, src: Column) -> Column {
+    fn replace_col(&mut self, c: usize, src: Column) -> Column {
         mem::replace(&mut (*self)[c], src)
     }
 
     /// Get a row from this array by-value.
-    fn row(&self, r: uint) -> Row;
+    fn row(&self, r: usize) -> Row;
 
     /// Swap two rows of this array.
-    fn swap_rows(&mut self, a: uint, b: uint);
+    fn swap_rows(&mut self, a: usize, b: usize);
 
     /// Swap the values at index `a` and `b`
     #[inline]
-    fn swap_elems(&mut self, a: (uint, uint), b: (uint, uint)) {
+    fn swap_elems(&mut self, a: (usize, usize), b: (usize, usize)) {
         let (ac, ar) = a;
         let (bc, br) = b;
         unsafe { ptr::swap(&mut (*self)[ac][ar], &mut (*self)[bc][br]) };
